[refactor] 컨트롤러 패키지 구성 수정 ( #3 )
- UserController 와 HomeController 의 패키지 구성이 서로 바뀌어 있어서, 이를 개선함
