+++ src/main/java/com/coupon/issuecouponservice/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.coupon.issuecouponservice.controller.user;
+package com.coupon.issuecouponservice.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
